Replaced integer type with assured-size integer type.

+++ cliff.h
@@ -12,7 +12,7 @@
 typedef struct cliff cliff;
 
 /// Create and (possibly) attach a new command.
-int cliff_cmd(
+uint8_t cliff_cmd(
   cliff* cli,          ///< Command-line interface.
   uint64_t* cid,       ///< New returned command ID.
   const uint64_t pcid, ///< Parent command ID to attach to (0 for no parent).
@@ -23,7 +23,7 @@ int cliff_cmd(
 );
 
 /// Create and attach a new option.
-int cliff_opt(
+uint8_t cliff_opt(
   cliff* cli,          ///< Command-line interface.
   const uint64_t cid,  ///< Command ID to attach to.
   const char shtn,     ///< Short name (\0 for no short name).
@@ -35,7 +35,7 @@ int cliff_opt(
 );
 
 /// Create and attach a new flag.
-int cliff_flg(
+uint8_t cliff_flg(
   cliff* cli,          ///< Command-line interface.
   const uint64_t cid,  ///< Command ID to attach to.
   const char shtn,     ///< Short name (\0 for no short name).
@@ -46,7 +46,7 @@ int cliff_flg(
 );
 
 /// Create and attach a new argument.
-int cliff_arg(
+uint8_t cliff_arg(
   cliff* cli,          ///< Command-line interface.
   const uint64_t cid,  ///< Command ID to attach to.
   const char* name,    ///< Name.
@@ -55,7 +55,7 @@ int cliff_arg(
 );
 
 /// Skip a number of bytes in the storage structure.
-int cliff_pad(
+uint8_t cliff_pad(
   cliff* cli,          ///< Command-line interface.
   const uint64_t cid,  ///< Command ID.
   const size_t size    ///< Number of bytes to skip.
@@ -63,7 +63,7 @@ int cliff_pad(
 
 /// Parse the command-line arguments. This function frees all memory allocated
 /// the build functions.
-int cliff_run(
+uint8_t cliff_run(
   cliff* cli,          ///< Command-line interface.
   const uint64_t rcid, ///< Root command ID (name used as program name).
   const int argc,      ///< Argument count.
@@ -74,6 +74,6 @@ int cliff_run(
 /// Convert an error code to an English description.
 void cliff_err(
   cliff* cli,         ///< Command-line interface.
-  const int code,     ///< Function return/error code.
+  const uint8_t code, ///< Function return/error code.
   const char** estr   ///< English description.
 );
